requirements.txt
streamlit
nltk
scikit-learn
pandas
numpy
fasttext-0.9.2-cp311-cp311-win_amd64.whl